requirements.txt
opencv-python
tensorflow
pillow
numpy